pyproject.toml
[project]
name = "hotel-lakatsiya"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-stubs>=3.16.7",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
]
